Restructure code for angelic conditions
* For now, angelic execution is bundled with the rest. In the future, consider moving it to HerbInterpret
