+++ src/frangel/frangel_generation.jl
@@ -227,42 +227,3 @@ function get_descendant_replacements!(node::RuleNode, symbol::Symbol, grammar::A
         get_descendant_replacements!(child, symbol, grammar, replacements)
     end
 end
-
-"""
-    add_angelic_conditions!(program::RuleNode, grammar::AbstractGrammar, angelic_conditions::Dict{UInt16,UInt8})::RuleNode
-
-Add angelic conditions to a program. This is done by replacing some of the nodes indicated by `angelic_conditions`` with holes.
-
-# Arguments
-- `program`: The program to modify.
-- `grammar`: The grammar rules of the program.
-- `angelic_conditions`: A dictionary mapping indices of angelic condition candidates, to the child index that may be changed.
-
-# Returns
-The modified program with angelic conditions added.
-
-"""
-function add_angelic_conditions!(program::RuleNode, grammar::AbstractGrammar, angelic_conditions::Dict{UInt16,UInt8})::RuleNode
-    if isterminal(grammar, program.ind)
-        return program
-    end
-    # If the current node has an angelic child, look for it
-    if haskey(angelic_conditions, program.ind)
-        angelic_condition_ind = angelic_conditions[program.ind]
-        for (index, child) in enumerate(program.children)
-            # Traverse children for angelic condition candidates
-            if index != angelic_condition_ind
-                program.children[index] = add_angelic_conditions!(child, grammar, angelic_conditions)
-                # A hole represents the angelic condition's location - to be replaced by angelic rulenode before evaluation
-            else
-                program.children[index] = Hole(grammar.domains[grammar.childtypes[program.ind][angelic_condition_ind]])
-            end
-        end
-        # Traverse the node's children for angelic condition candidates
-    else
-        for (index, child) in enumerate(program.children)
-            program.children[index] = add_angelic_conditions!(child, grammar, angelic_conditions)
-        end
-    end
-    program
-end
